requirements.txt
asgiref==3.5.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-googledrive-storage==1.6.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
Pillow==9.1.0
protobuf==3.20.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
reportlab==3.6.9
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
